Count download size when downloading instead of after
