+++ src/toil/jobStores/googleJobStore.py
@@ -349,6 +349,7 @@ class GoogleJobStore(AbstractJobStore):
     def _readFromUrl(cls, url, writable):
         blob = cls._getBlobFromURL(url, exists=True)
         blob.download_to_file(writable)
+        return blob.size
 
     @classmethod
     def _supportsUrl(cls, url, export=False):
